Main guard now included in primary group by dask file
